requirements.txt
contourpy==1.3.2
cycler==0.12.1
fonttools==4.62.1
greenlet==3.4.0
kiwisolver==1.5.0
matplotlib==3.10.8
numpy==2.2.6
packaging==26.0
pandas==2.3.3
pillow==12.2.0
psycopg2-binary==2.9.11
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
six==1.17.0
SQLAlchemy==2.0.49
typing_extensions==4.15.0
tzdata==2026.1
